+++ src/Fields/ID.php
@@ -115,6 +115,14 @@ class ID implements IDContract, Fillable
     /**
      * @inheritDoc
      */
+    public function mustExist(): bool
+    {
+        return false;
+    }
+
+    /**
+     * @inheritDoc
+     */
     public function fill(Model $model, $value, array $validatedData): void
     {
         $column = $this->column() ?: $model->getRouteKeyName();