+++ tests/lib/Integration/Fields/ArrayListTest.php
@@ -20,6 +20,7 @@ declare(strict_types=1);
 namespace LaravelJsonApi\Eloquent\Tests\Integration\Fields;
 
 use App\Models\Role;
+use App\Models\User;
 use Illuminate\Http\Request;
 use LaravelJsonApi\Eloquent\Fields\ArrayList;
 use LaravelJsonApi\Eloquent\Tests\Integration\TestCase;
@@ -95,7 +96,9 @@ class ArrayListTest extends TestCase
         $model = new Role();
         $attr = ArrayList::make('permissions');
 
-        $attr->fill($model, $value, []);
+        $result = $attr->fill($model, $value, []);
+
+        $this->assertNull($result);
         $this->assertSame($value, $model->permissions);
     }
 
@@ -180,6 +183,17 @@ class ArrayListTest extends TestCase
 
         $attr->fill($model, ['foo', 'bar'], []);
         $this->assertSame(['bar', 'foo'], $model->permissions);
+    }
+
+    public function testFillRelated(): void
+    {
+        $user = new User();
+
+        $attr = ArrayList::make('permissions')->on('profile')->unguarded();
+
+        $attr->fill($user, ['foo', 'bar'], []);
+
+        $this->assertEquals(['foo', 'bar'], $user->profile->permissions);
     }
 
     public function testReadOnly(): void
@@ -283,6 +297,19 @@ class ArrayListTest extends TestCase
         $attr = ArrayList::make('permissions')->sorted();
 
         $this->assertSame(['bar', 'foo'], $attr->serialize($model));
+    }
+
+    public function testSerializeRelated(): void
+    {
+        $user = new User();
+
+        $attr = ArrayList::make('permissions')->on('profile');
+
+        $this->assertNull($attr->serialize($user));
+
+        $user->profile->permissions = ['foo', 'bar'];
+
+        $this->assertEquals(['foo', 'bar'], $attr->serialize($user));
     }
 
     public function testHidden(): void