+++ tests/lib/Integration/Fields/BooleanTest.php
@@ -97,7 +97,9 @@ class BooleanTest extends TestCase
         $model = new User();
         $attr = Boolean::make('admin');
 
-        $attr->fill($model, $value, []);
+        $result = $attr->fill($model, $value, []);
+
+        $this->assertNull($result);
         $this->assertSame($value, $model->admin);
     }
 
@@ -170,6 +172,17 @@ class BooleanTest extends TestCase
 
         $attr->fill($user, true, []);
         $this->assertFalse($user->admin);
+    }
+
+    public function testFillRelated(): void
+    {
+        $user = new User();
+
+        $attr = Boolean::make('admin')->on('profile')->unguarded();
+
+        $attr->fill($user, true, []);
+
+        $this->assertTrue($user->profile->admin);
     }
 
     public function testReadOnly(): void
@@ -261,6 +274,19 @@ class BooleanTest extends TestCase
         });
 
         $this->assertFalse($attr->serialize($model));
+    }
+
+    public function testSerializeRelated(): void
+    {
+        $user = new User();
+
+        $attr = Boolean::make('admin')->on('profile');
+
+        $this->assertNull($attr->serialize($user));
+
+        $user->profile->admin = true;
+
+        $this->assertTrue($attr->serialize($user));
     }
 
     public function testHidden(): void