+++ tests/lib/Integration/Fields/DateTimeTest.php
@@ -20,7 +20,9 @@ declare(strict_types=1);
 namespace LaravelJsonApi\Eloquent\Tests\Integration\Fields;
 
 use App\Models\Post;
+use App\Models\User;
 use Carbon\Carbon;
+use Carbon\Traits\Creator;
 use Illuminate\Http\Request;
 use LaravelJsonApi\Eloquent\Fields\DateTime;
 use LaravelJsonApi\Eloquent\Tests\Integration\TestCase;
@@ -96,7 +98,9 @@ class DateTimeTest extends TestCase
 
         $attr = DateTime::make('publishedAt')->retainTimezone();
 
-        $attr->fill($model, null, []);
+        $result = $attr->fill($model, null, []);
+
+        $this->assertNull($result);
     }
 
     public function testAppTimezone(): void
@@ -219,6 +223,17 @@ class DateTimeTest extends TestCase
         $this->assertSame('2020-11-22T11:48:17.000000Z', $post->published_at->toJSON());
     }
 
+    public function testFillRelated(): void
+    {
+        $user = new User();
+
+        $attr = DateTime::make('publishedAt')->on('profile')->unguarded();
+
+        $attr->fill($user, '2020-11-23T11:48:17.000000Z', []);
+
+        $this->assertEquals(Carbon::parse('2020-11-23T11:48:17.000000Z'), $user->profile->published_at);
+    }
+
     public function testReadOnly(): void
     {
         $request = $this->createMock(Request::class);
@@ -296,6 +311,19 @@ class DateTimeTest extends TestCase
         });
 
         $this->assertEquals(new Carbon('2020-02-01 00:00:00'), $attr->serialize($model));
+    }
+
+    public function testSerializeRelated(): void
+    {
+        $user = new User();
+
+        $attr = DateTime::make('createdAt')->on('profile');
+
+        $this->assertNull($attr->serialize($user));
+
+        $user->profile->created_at = $expected = new Carbon('2020-02-01 15:55:00');
+
+        $this->assertEquals($expected, $attr->serialize($user));
     }
 
     public function testHidden(): void